requirements.txt
#Testing dependencies
pytest==8.3.5
pytest-cov==6.1.1
factory-boy==3.3.3

#Code quality
pylint==3.3.6
flake8==7.2.0

#Pantry dependencies
pymongo==4.12.1